fix: add lodash to devDependencies for fixing a security issue
Changes in this commit must be reverted in futuer.

The following packages in devDependencies depend on lodash:

* nyc

and causes the following security issue:

* [CVE-2019-10744]

The security issue has already fixed in lodash, but these modules use a old
version at this point even when I update packages with `npm update`.

lodash has been added to devDependencies in order to fix the security issue even
though no function in lodash is directly used in tests.

[CVE-2019-10744]: lodash/lodash#4336